+++ libs/contract-nft/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@thecointech/contract-tools/jest.config');