+++ libs/nft-contract/migrations/3_dev_initialize.js
@@ -1,1 +1,0 @@
-module.exports = require(`./3_dev_initialize.ts`)(artifacts, web3);