+++ libs/site-base/internal/server/port.mjs
@@ -1,0 +1,3 @@
+import args from './argv.mjs';
+
+export const port = parseInt(args.port || process.env.PORT || '3001', 10);